cmake: switch from FindElektra.cmake to ElektraConfig.cmake
Instead of providing a FindElektra.cmake module to find an installed
elektra, provide a ElektraConfig.cmake module with the configuration
of the installed elektra. This has few advantages:
- ElektraConfig.cmake imports the targets and sets variables for paths,
  since they are known when building/installing elektra
- the configuration files are not installed among the private CMake
  module path

The additional generated ElektraConfigVersion.cmake file takes care of
the version handling, so specifying a version in
  find_package(Elektra <version>)
now will do the proper thing.

Following CMake' style, now exported variables are Elektra_*, with
ELEKTRA_* provided for compatibility.