python: fix flags in Key()
One of the constructors of Key takes uint64_t, coming from what swig
maps for an int value. OTOH, passing a uint64_t as vararg value for
KEY_FLAGS will lead to mismatches later, as KEY_FLAGS is read as int.

Hence, check that the value passed in the uint64_t can fit into an int,
and cast it as such when passing it to KEY_FLAGS.