Do not create directories `(` and `)` inside `doc`
Before this change CMake would create the folders `(` and `)` inside
`doc`. This was caused by the following command:

    file(MAKE_DIRECTORY(${folder}))

The command `file(MAKE_DIRECTORY [<directories>...])` creates the given
directories if the do not already exist. If we use the arguments
`(${folder})` after the argument `MAKE_DIRECTORY`, then the `file`
command interprets these as the list containing `(`, `${folder}` and
`)`. Therefore the command would also create the directories `(` and
`)`, although we only wanted to make sure that the folder specified in
the variable `folder` exists.

This commit fixes #438.