Fix documentation of `elektrify-getenv`
Before this commit, the documentation stated that the command `kdb
elektrify-getenv` can be used to change the value of environment
variables. This however does not work. Basically `kdb` interprets
`elektrify-getenv` as – non existent – subcommand and displays an error
message after we enter the following in a shell:

    kdb elektrify-getenv man man --elektra:MANWIDTH=40

To fix this, use the following command instead:

    elektrify-getenv man man --elektra:MANWIDTH=40

After you entered the text above, `man` displays its own man page using
a maximum line length of 40 characters.