Add redrawn version of logo
This commit contains new slightly modified versions of the Elektra
logo. We replace each occurrence of the old logos, inside the
documentation and the source tree, with the new version.

There are now three versions of the Elektra logo:

  - Circle
  - Circle with shadows
  - Circle with shadows and background

We save each of this files as SVG. If you want to create a bitmap
version of one of the logos, we recommend `svgexport` [1]. For example,
we created `elektra-logo.png` inside `src/tools/qt-gui/qml/icons/` via
the following command:

    svgexport circle_with_shadows.svg elektra-logo.png 100:100

[1]: https://github.com/shakiba/svgexport