Tests: Remove useless CMake linking code
We introduced the linking code to handle problems with GCC version 5.3
[#432] and modified it to work with static linking only [#489].
However, it looks like we do not need it anymore. Building on (OS X)
works without the code. For example, the commands

  cmake -DBUILD_DOCUMENTATION=OFF -DBUILD_FULL=OFF -DBUILD_PDF=OFF \
  -DBUILD_SHARED=OFF -DBUILD_STATIC=ON ..
  make -j 9

works with Clang version `clang-703.0.29` and GCC version `5.3.0`.
The same can be said if we build Elektra using the commands

  cmake ..
  make -j 9

.

[#432]: https://github.com/ElektraInitiative/libelektra/issues/432
[#489]: https://github.com/ElektraInitiative/libelektra/issues/489