Libease: Fix possible problem with cascading keys
Before this change `elektraKeyGetRelativeName` assumed that the prefix
of a cascading key is a single level name. So a possible child key of
`/cascading` is `system/cascading/key` or `user/cascading/key`, but not
`user/second_level/cascading/key`. This update makes sure that the
function also works if there are multiple levels of names before the
cascading part of the child's name.