Tests: Fix some test cases in resolver check
Testcases that create temporary directories would not work, if the
command `mktemp -d` returns a path containing a link. For example in OS
X the command returns something like `/var/folders/random_dir`, which in
reality is `private/var/folders/random_dir`, since `/var` is a link to
`/private/var`.