Documentation: Add accent to name
We had problems with the character `é` (small letter e + combining
acute accent — code points 101 and 769) before. LaTeX – with input
encoding set to `utf8` – is not able to use this character. Because of
this problem, we could not translate the reference manual. We fix this
by using `é` (small letter e with acute — code point 233) instead of
`é`.

This fixes #636.