Qt GUI xdg icon theme support rework:
Rename desktop file: correct reverse url from org.elektra to org.libelektra. Rename elektra-qt to elektra-qt-editor and use that in future references to the qt-gui as it is a clearer name about the function of the program and binary.

Rename ChooseColorWindow: The ChooseColorWindows will be replaced by a
AppearanceSettingsWindow, all references to ChooseColor, choose color have been
replaced by AppearanceSettings or choose appearance.