Qt GUI xdg icon theme support rework:
Introduce Appearance Settings Window: Appearance Settings Window
contains both color settings as well as a switch to disable or enable
the system icon theme. For this to work we had to introduce the setting
in `guisettings`. We also added a private function in `guisettings` to
get and set settings with a boolean value.

Tree reload on Settings close: We now synch and refresh the tree view on
closing of the settings window if any settings have been changed, so
changes can be seen imediatly in the tree.