Elektra Qt Editor: save settings on close signal
Currently on any close signal the Color/Appearance Settings Dialog did
not save any options when on the close signal, but did not reset any set
options either.

With this change the dialog reacts the same way on closing signals as it
did by pressing the offered close button by saving the changes.

The alternative would be to over a ok button and a cancel button. And in
consequence reset settings on close and cancel signals, and only save
them when the ok button is pressed.