ASAN: correct libasan linking (gcc vs clang), workaround gtest gcc
We should only pass -lasan when using gcc and ASAN_ENABLED as this
otherwise results in clang linking the gcc sanitizer which in return
results in broken binarys.

Also currently use a workaround to add -lpthread when using ASAN with
gcc. This probably should be moved into the GTest logic of the cmake
files, but i am currently lacking the expertise to do this correctly.