Build: Add spec as dependency of `kdberrors.h`
We add a dependency for the specification file to the custom command
that generates `kdberrors.h`. This way the build system will regenerate
`kdberrors.h` every time we update the specification file.

Before this change `kdberrors.h` was only updated if the file
`kdberrors.h` did not exist beforehand.

This commit fixes #893.