Shell Recorder: Fix broken protocol files
Before this change Shell Recorder would not print newlines after it
wrote a message to the protocol file. This commit fixes this regression
introduced in commit 5d2232ed.

The problem was that we replaced `echo -e` with `printf` but did not
consider that `printf` does not print a newline at the end of a message.

This commit fixes #903.