Ruby: Ruby module code kdb.rb (to_s, pretty_print)
This adds the old and removed kdb.rb main module again. However, we've
fixed the problems in the first place. Now we tell Swig to create a
'kdb' module (instead '_kdb') but name the 'Init' function '_kdb'. This
allows us to name the C-extension library '_kdb.so'. So, a
"require '_kdb'" will load this extension, which defines the 'kdb'
module. Now we can add a 'kdb.rb' Ruby file, which loads the C-extension
library and extends the 'Kdb' module and its classes (Ruby allows to
extend existing classes, without polluting inheritance hierarchies).
This way the user will just use "require 'kdb'", which loads the
"kdb.rb" file, which loads the C-extension library.

This methods allows to extend the binding by Ruby code, which makes
development much easier.

First extension:
 - Key.to_s, Key.pretty_print
 - KeySet.to_s, KeySet.pretty_print