Completions: Generalize function
We now use the function `__input_includes` to check if the current
commandline buffer contains at least one value of a given list.

This function is useful, since there are many situations where the Fish
completion commands need to provide different completions if the
command line buffer contains a certain word. For example, it does not
make sense for Fish to suggest a subcommand if the current command line
buffer already contains another subcommand.