Shell Recorder: Fix environment variable `PATH`
Before this change `shell_recorder` would prefer commands inside the
build folder over the ones on the system. As a consequence of this the
shell recorder would call the command `cut` inside the build folder
instead of the one shipped with the system. This is a problem since the
program `cut` inside the build folder is clearly not a version of the
Unix utility `cut`.