build-homepage: late fixes
Do we really want to run the backend service as an ASAN build? It
already crashed two times because of that.

Remove installed libraries before installing the new one. This way we
ensure there are no old unwanted plugins that might cause problems.

Ensure rest-backend is started in daemon mode, otherwise the script will
be blocked.