Test: Do not close modules twice
Before this change the test would cause a segmentation fault if the
default storage plugin specified a function for `ELEKTRA_PLUGIN_CLOSE`.
This was the case since both `elektraModulesClose` and `kdb_del` close
the default storage plugin. Under this circumstances `kdb_del` would
try to close an already closed plugin and fail.