Clang-Format: Add NL after `union`
We now explicitly add a new line after the keyword `union`, but not
after the keyword `enum`. In theory we should add newlines after `enum`
too. However, since Clang-Format 5.0 does not detect as many enum
definitions as Clang-Format 7.0 we do not add a newline. This has the
advantage, that the formatted output of Clang-Format 5.0 and
Clang-Format 7.0 is identical.