haskell:Check GHC version in cmake
The Haskell binding requires haskell base>=4.7 to be available on the
system. This is provided by GHC if it's version is at least 7.8.1.

However it was decided to go with a required version of GHC 8.0.0 as
per discussion in the issue mentioned below.

The introduced checks require GHC with an appropriate version to be
installed if haskell bindings are supposed to be build.

Fixes #1891