Shell Recorder: Print content of protocol file
The Shell Recorder now also prints the content of the protocol file if
a test was unsuccessful or you used the command switch `-p`. This
update makes sure, that you can always check the standard error output
of individual test commands, if a Shell Recorder test failed.

This commit closes #1936.