Release: Improve structure of release notes
We now use a structure, similar to the one of the repository, for the
release notes. Instead of personal interest (maintainer, developer,
user) we now use general technical topics (bindings, plugins, tools, …)
as headings. This structure should it make much simpler to decide where
to look for a certain new feature or bug fix.