core/cache: move PREGETSTORAGE after cache hit check, skip default split on POSTGETSTORAGE
We don't want any plugins to be executed on cache hits. Therefore we
move the PREGETSTORAGE position after the cache hit check. Additionally
we may not cache KeySets which are not from backends. Thus we skip the
default split on the POSTGETSTORAGE position, which allows us to cache
the completely finished KeySet and not call POSTGETSTORAGE each time.