Reformat CMake: Fix version detection
The command `cmake-format --version` seems to print the version number
either to `stdout` or `stderr`, depending on the current system. To
make sure we always capture the version number correctly we therefore
redirect `stderr` to the same file descriptor as `stdout`, before we
save the output of the command.

This commit closes #2874.