Fix mmapstorage test
To always ksDel() the key->meta keyset and
recreate it in keyInit makes the
test_mmapstorage fail. It works fine for the abi,
but since the mmapstorage plugin uses
the keyset and key fields directly, it is not
clear where exactly that results in an error.

This change is both abi compatible and makes
mmapstorage happy.