toml: Added wr tests for comments. Fixed comment error.
First newline in file was not counted, since newlines rules in grammar don't count first newline (since the rule is expected to be applied after a node, where theres always a newline. Added separate rule for the leading newlines, where all newlines are counted. Also removed extra spacing after/before array/inline table braces, because was introducing extra whitespace to comments on read (but it's only relevant for arrays anyway). Need a check for a array leading comment to fix this properly.