toml: Added doc, fixed possible ordering messup
Ordering has now an additional constraint for elements under the same TOML hierarchial element (eg table/table array). Table array and table keys are now always sorted after non-table keys, so that keys not belonging to a table will not be placed after a table declaration (which would make them a member of that table on a subsequent read). Check README.md on ordering for details.