toml: Fixed writing of incomplete comment info.
The plugin now fills in missing comment metakeys (start/spaces) when writing.
Previously, if a comment was created via kdb meta-set, it would miss the
start and space metakeys, resulting in the comment written without it's
starting symbol, creating invalid TOML files.

Fixed an infinte loop, which occurec when the root key was in the
keyset.

Fixed a test case, which didn't test file ending comments properly.
Added some doc for comments.