toml: Fixed error in table array comments.
Fixes #3474
The problem was, that when having comments associated to
table array declarations in a TOML file, the table element root key (eg.
tablearray/#0), which normally is not emitted in the keyset, has to be added to it.
This caused trouble in the building of the file tree for writing.
Fixed the table array index loop, to handle the element root key
appropriatly.