toml: Fixes on writing of scalar values.
Added better handling when a `type` metakey is present.
Before, when a `type` metakey was present, the value would be written
as-is, assuming it's a number. However, when the `type` metakey was eg.
set to `wstring`, the value being definitly not a number, it would also write it as-is,
producing possible bare-strings on the rhs, which is not valid.
Will now only write unquoted values if it's a number/date/boolean,
reqardless of the `type` metakey.