toml: Added basic metakey preservation.
Metakeys are now written as comments with a prefix on writing.
On reading, those special comments will get assigned as metakeys to a
key.

This is currently just a very basic implementation, which assumes that
the metakey name/values don't have any special characters / are binary.
Also, metakey names are currently expected to not contain any spaces.

Tests for this changes not yet added/adapted.