doc: Make tutorials/namespaces verifiable
The tutorial is pernicious because it teaches how to modify the
configuration of Elektra itself using Elektra. Since it is undesired for
the tutorial verificication to modify the configuration, these commands
have been excluded from the shell recorder verification by changing the
code fence from `\`\`\`sh` to `\`\`\`shell`.

Furthermore, the tutorial previously featured the creation of keys
without specifying a namespace. This is no longer supported and the
tutorial has been update accordingly.

Since one of the images was a bit misleading it has been removed.